+++ backend/src/main/java/com/team5/secondhand/chat/bubble/controller/ChatController.java
@@ -1,15 +1,17 @@
-package com.team5.secondhand.chat.controller;
+package com.team5.secondhand.chat.bubble.controller;
 
-import com.team5.secondhand.chat.domain.ChatBubble;
-import com.team5.secondhand.chat.service.RedisMessagePublisher;
+import com.team5.secondhand.chat.bubble.domain.ChatBubble;
+import com.team5.secondhand.chat.bubble.service.RedisMessagePublisher;
+import com.team5.secondhand.chat.bubble.service.RedisMessageSubscriber;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
-@Controller
+@RestController
 @Slf4j
 class ChatController {
     private final RedisMessagePublisher redisMessagePublisher;
@@ -20,4 +22,5 @@ class ChatController {
         log.debug("pub controller");
         redisMessagePublisher.publish(channelTopic.getTopic(), message);
     }
+
 }
