+++ backend/src/main/java/com/team5/secondhand/chat/bubble/service/RedisMessagePublisher.java
@@ -1,6 +1,7 @@
-package com.team5.secondhand.chat.service;
+package com.team5.secondhand.chat.bubble.service;
 
-import com.team5.secondhand.chat.domain.ChatBubble;
+import com.team5.secondhand.chat.bubble.domain.ChatBubble;
+import com.team5.secondhand.chat.bubble.repository.ChatBubbleRepository;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,9 +14,11 @@ import org.springframework.stereotype.Service;
 @Service
 public class RedisMessagePublisher {
     private final RedisTemplate<String, Object> redisTemplate;
+    private final ChatBubbleRepository chatBubbleRepository;
 
     public void publish(String topic, ChatBubble message) {
         log.debug("pub log : " +  message.toString() + "/ topic: " + topic);
+        chatBubbleRepository.save(message);
         redisTemplate.convertAndSend(topic, message);
     }
 }
