--- backend/src/main/java/com/team5/secondhand/chat/domain/MessageType.java
@@ -1,5 +1,0 @@
-package com.team5.secondhand.chat.domain;
-
-public enum MessageType {
-    ENTER, TEXT;
-}
