--- backend/src/main/java/com/team5/secondhand/chat/repository/ChatRoomRepository.java
@@ -1,11 +1,0 @@
-package com.team5.secondhand.chat.repository;
-
-import com.team5.secondhand.chat.domain.ChatRoom;
-
-import java.util.*;
-
-public interface ChatRoomRepository {
-    List<ChatRoom> findAll();
-    Optional<ChatRoom> findById(String roomId);
-    boolean save(ChatRoom chatRoom);
-}
